Added ability to draw lines, which are saved in arraylist

+++ Decorator/src/backend/Circle.java
@@ -3,32 +3,32 @@ package backend;
 public class Circle {
 
 	private final double diameter = 4.0;
-	private double x;
-	private double y;
+	private int x;
+	private int y;
 	
 	/**
 	 * @param x the x coordinate location
 	 * @param y the y coordinate location
 	 */
-	public Circle(double x, double y) {
+	public Circle(int x, int y) {
 		super();
 		this.x = x;
 		this.y = y;
 	}
 
-	public double getX() {
+	public int getX() {
 		return x;
 	}
 
-	public void setX(double x) {
+	public void setX(int x) {
 		this.x = x;
 	}
 
-	public double getY() {
+	public int getY() {
 		return y;
 	}
 
-	public void setY(double y) {
+	public void setY(int y) {
 		this.y = y;
 	}
